app/api/requirements.txt
flask
langchain-google-genai
pydantic
python-dotenv
langchain-openai
browser-use
playwright install